+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 
 - `alias` for aliasing boxes
+- `env` for accessing values from the env
 
 ## [0.1.0] - 2017-01-29
 ### Added